pyproject.toml
[tool.poetry]
name = "dialo-clone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dialo_clone"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.100.0"}
umongo = "^3.1.0"
motor = "^3.2.0"
python-decouple = "^3.8"
pyjwt = "^2.7.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
